refactor(domain): 야구공 비교 관련 로직 변경
기존 Ball에서 자리수를 제거하고 BallNumber로 이름 변경
비교 로직을 Balls가 갖도록 수정
